Create BCFTools_View_Concat_Prune
The first command in this file (bcftools view) creates vcfs from all BCFs after mpileup. Because each part has the same order of individuals, with slightly different names, we use "grep -v \#" to take the names off of parts 1-12. These BCFs are then concatenated together using cat, which brings together all the files using the naming convention of part1. After this, bcftools view is used to filter the data and finally bcftools prune is used to prune sites for LD and missingness.